requirements.txt
pydantic-ai
python-dotenv 